refactor: text input with function component & useForm
