+++ src/components/Form/Inputs/TextInput.tsx
@@ -1,30 +1,31 @@
 import React from 'react';
+import { FieldValues, UseFormRegister } from 'react-hook-form';
 
 interface IProps {
   label: string;
-  reference: React.RefObject<HTMLInputElement>;
-  correct: boolean;
+  name: string;
+  reference: UseFormRegister<FieldValues>;
+  error?: string;
 }
 
-class TextInput extends React.Component<IProps> {
-  constructor(props: IProps) {
-    super(props);
-  }
-  render() {
-    return (
-      <fieldset>
-        <label className="form__label form__label-block">
-          {this.props.label} {this.props.correct && <span style={{ color: 'red' }}>incorrect</span>}
-        </label>
-
-        <input
-          className="form__input_text"
-          type="text"
-          placeholder={`Input ${this.props.label}`}
-          ref={this.props.reference}
-        />
-      </fieldset>
-    );
-  }
+export default function TextInput(props: IProps) {
+  return (
+    <fieldset>
+      <label className="form__label form__label-block">{props.label}</label>
+      <input
+        className="form__input_text"
+        type="text"
+        placeholder={`Input ${props.label}`}
+        autoComplete="off"
+        {...props.reference(props.name, {
+          required: true,
+          minLength: {
+            value: 3,
+            message: 'Title is too short, minimum three letters',
+          },
+        })}
+      />
+      {props.error && <span style={{ color: 'red' }}>{props.error}</span>}
+    </fieldset>
+  );
 }
-export default TextInput;
